Added support for Ubuntu Bionic
Bionic uses Gnome instead of Unity.
