+++ molecule/gnome/molecule.yml
@@ -9,7 +9,7 @@ lint:
   name: yamllint
 
 platforms:
-  - name: ansible-role-xdesktop-ubuntu-bionic-default
+  - name: ansible-role-xdesktop-ubuntu-bionic-gnome
     image: ubuntu:18.04
 
 provisioner:
@@ -18,7 +18,7 @@ provisioner:
     name: ansible-lint
 
 scenario:
-  name: default
+  name: gnome
 
 verifier:
   name: testinfra
