+++ tasks/gnome-desktop.yml
@@ -1,4 +1,4 @@
 ---
-- import_tasks: unity-desktop-install.yml
+- import_tasks: gnome-desktop-install.yml
 
 - import_tasks: gnome-desktop-configure.yml
