+++ tasks/main.yml
@@ -2,7 +2,7 @@
 - name: assert desktop environment in supported list
   assert:
     that:
-      - "xdesktop_desktop in ('unity', 'xfce4')"
+      - "xdesktop_desktop in ('gnome', 'unity', 'xfce4')"
 
 - include_tasks: '{{ xdesktop_desktop }}-desktop.yml'
